requirements.txt
git+https://github.com/sunsmarterjie/yolov12.git
torch
torchvision
streamlit
Pillow
opencv-python-headless
